Removed duplicate class StaticUnknownPacketFactory in pcap4j-packetfactory-static.
